refactor: 관련 로직 함수 근접한 위치에 오도록 수정
